backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-caching==2.1.0
feedparser==6.0.10
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
